requirements.txt
numpy
pillow
torch
torchvision
matplotlib
scikit-learn
